Updated readme, converted AllStudents to class component
