+++ client/components/StudentCard.jsx
@@ -2,7 +2,10 @@ import React from 'react';
 import propTypes from 'prop-types';
 
 
-// Set up some sort of component that pops up with student info after picking student
+// TO DO
+// Wrap student card in div, move button to outside div and change to home icon
+// Center student card?
+// Time to start prettying up this app
 
 const StudentCard = ({ onClose, data: { name, profilePic, lastCalled, timesCalled }  }) => (
   <div className = 'student-card'>
